Function amazon_sp_easy_ship::apis::easy_ship_api::create_scheduled_package_bulk

pub async fn create_scheduled_package_bulk(
    configuration: &Configuration,
    create_scheduled_packages_request: CreateScheduledPackagesRequest
) -> Result<CreateScheduledPackagesResponse, Error>

This operation automatically schedules a time slot for all the amazonOrderIds given as input, generating the associated shipping labels, along with other compliance documents according to the marketplace (refer to the marketplace document support table). Developers calling this operation may optionally assign a packageDetails object, allowing them to input a preferred time slot for each order in ther request. In this case, Amazon will try to schedule the respective packages using their optional settings. On the other hand, i.e., if the time slot is not provided, Amazon will then pick the earliest time slot possible. Regarding the shipping label’s file format, external developers are able to choose between PDF or ZPL, and Amazon will create the label accordingly. This operation returns an array composed of the scheduled packages, and a short-lived URL pointing to a zip file containing the generated shipping labels and the other documents enabled for your marketplace. If at least an order couldn’t be scheduled, then Amazon adds the rejectedOrders list into the response, which contains an entry for each order we couldn’t process. Each entry is composed of an error message describing the reason of the failure, so that sellers can take action. The table below displays the supported request and burst maximum rates: | Rate (requests per second) | Burst | | –– | –– | | 1 | 5 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.